refactor the forms.py file
