+++ tasks/forms/__init__.py
@@ -1,0 +1,6 @@
+from .form_mixins import *
+from .account_forms import *
+from .authenticate_forms import *
+from .avatar_form import *
+from .team_form import *
+from .upload_form import *
