+++ tasks/forms/authenticate_forms.py
@@ -1,0 +1,50 @@
+"""Authentication Forms for the app."""
+from django import forms
+from django.contrib.auth import authenticate
+from tasks.models import User
+from . import NewPasswordMixin
+
+
+class LogInForm(forms.Form):
+    """Form enabling registered users to log in."""
+
+    username = forms.CharField(label="Username", label_suffix='')
+    password = forms.CharField(label="Password", widget=forms.PasswordInput(), label_suffix='')
+
+    def get_user(self):
+        """Returns authenticated user if possible."""
+
+        user = None
+        if self.is_valid():
+            username = self.cleaned_data.get('username')
+            password = self.cleaned_data.get('password')
+            user = authenticate(username=username, password=password)
+        return user
+
+
+class SignUpForm(NewPasswordMixin, forms.ModelForm):
+    """Form enabling unregistered users to sign up."""
+
+    class Meta:
+        """Form options."""
+
+        model = User
+        fields = ['first_name', 'last_name', 'username', 'email']
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for field in self.fields:
+            self.fields[field].label_suffix = ''
+
+    def save(self):
+        """Create a new user."""
+
+        super().save(commit=False)
+        user = User.objects.create_user(
+            self.cleaned_data.get('username'),
+            first_name=self.cleaned_data.get('first_name'),
+            last_name=self.cleaned_data.get('last_name'),
+            email=self.cleaned_data.get('email'),
+            password=self.cleaned_data.get('new_password'),
+        )
+        return user
