+++ tasks/forms/form_mixins.py
@@ -1,0 +1,27 @@
+"""Form mixin for the app."""
+from django import forms
+from django.core.validators import RegexValidator
+
+
+class NewPasswordMixin(forms.Form):
+    """Form mixing for new_password and password_confirmation fields."""
+
+    new_password = forms.CharField(
+        label='Password',
+        widget=forms.PasswordInput(),
+        validators=[RegexValidator(
+            regex=r'^(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9]).*$',
+            message='Password must contain an uppercase character, a lowercase '
+                    'character and a number'
+        )]
+    )
+    password_confirmation = forms.CharField(label='Password confirmation', widget=forms.PasswordInput())
+
+    def clean(self):
+        """Form mixing for new_password and password_confirmation fields."""
+
+        super().clean()
+        new_password = self.cleaned_data.get('new_password')
+        password_confirmation = self.cleaned_data.get('password_confirmation')
+        if new_password != password_confirmation:
+            self.add_error('password_confirmation', 'Confirmation does not match password.')
